Improve tests and manage all tests in one file
