+++ test/perpetual/perpLemma.js
@@ -51,6 +51,37 @@ async function callStaticOpenPosition(clearingHouse, signer, baseTokenAddress, _
     })
     return openPositionParams
 }
+async function openPosition(clearingHouse, signer, baseTokenAddress, _isBaseToQuote, _isExactInput, _amount) {
+    await clearingHouse.connect(signer).openPosition({
+        baseToken: baseTokenAddress,
+        isBaseToQuote: _isBaseToQuote,
+        isExactInput: _isExactInput,
+        oppositeAmountBound: 0,
+        amount: _amount, // amount in usd
+        sqrtPriceLimitX96: 0,
+        deadline: ethers.constants.MaxUint256,
+        referralCode: ethers.constants.HashZero,
+    })
+}
+
+async function addLiquidity(clearingHouse, signer, baseTokenAddress, _baseAmount, _quoteAmount, _lT, _uT) {
+    await clearingHouse.connect(signer).addLiquidity({
+        baseToken: baseTokenAddress,
+        base: _baseAmount,
+        quote: _quoteAmount,
+        lowerTick: _lT,
+        upperTick: _uT,
+        minBase: 0,
+        minQuote: 0,
+        useTakerBalance: false,
+        deadline: ethers.constants.MaxUint256,
+    })
+}
+
+async function forwardTimestamp(clearingHouse, step) {
+    const now = await clearingHouse.getBlockTimestamp()
+    await clearingHouse.setBlockTimestamp(now.add(step))
+}
 
 describe("perpLemma", async function () {
     let defaultSigner, usdLemma, reBalancer, hasWETH, keeperGasReward, signer1, signer2, usdl2, longAddress;
@@ -115,6 +146,9 @@ describe("perpLemma", async function () {
         ], { initializer: 'initialize' });
         await perpLemma.connect(signer1).resetApprovals()
 
+        // base = usd
+        // quote = eth
+
         await mockedBaseAggregator.setLatestRoundData(0, parseUnits("0.01", collateralDecimals), 0, 0, 0)
         await mockedBaseAggregator2.setLatestRoundData(0, parseUnits("100", collateralDecimals), 0, 0, 0)
 
@@ -127,34 +161,6 @@ describe("perpLemma", async function () {
         await marketRegistry.addPool(baseToken2.address, 10000)
         await marketRegistry.setFeeRatio(baseToken.address, 10000)
         await marketRegistry.setFeeRatio(baseToken2.address, 10000)
-
-        // prepare collateral for maker
-        const makerCollateralAmount = parseUnits("1000000", collateralDecimals)
-        await collateral.mint(signer1.address, makerCollateralAmount)
-        await collateral.mint(signer2.address, makerCollateralAmount)
-        
-        const parsedAmount = parseUnits("100000", collateralDecimals)
-        await collateral.connect(signer1).approve(vault.address, ethers.constants.MaxUint256)
-        await collateral.connect(signer2).approve(vault.address, ethers.constants.MaxUint256)
-
-        await collateral.mint(longAddress.address, parseUnits("10000000000", collateralDecimals))
-        await collateral.connect(longAddress).approve(vault.address, ethers.constants.MaxUint256)
-        await vault.connect(longAddress).deposit(collateral.address, parseUnits("10000", collateralDecimals))
-
-        // Deposit into vault
-        // await vault.connect(signer1).deposit(collateral.address, parsedAmount)
-        await vault.connect(signer2).deposit(collateral.address, parsedAmount)
-        await clearingHouse.connect(signer2).addLiquidity({
-            baseToken: baseToken.address,
-            base: parseEther('10000'),
-            quote: parseEther('100'),
-            lowerTick: -887200, //50000,
-            upperTick: 887200, //50400,
-            minBase: 0,
-            minQuote: 0,
-            useTakerBalance: false,
-            deadline: ethers.constants.MaxUint256,
-        })
     })
     
     beforeEach(async function () {
@@ -164,424 +170,728 @@ describe("perpLemma", async function () {
     afterEach(async function () {
         await revertToSnapshot(snapshotId);
     });
-    
-    
-    it("should set addresses correctly", async function () {
-        //setUSDLemma
-        await expect(perpLemma.connect(signer1).setUSDLemma(signer1.address)).to.be.revertedWith("Ownable: caller is not the owner");
-        await perpLemma.connect(defaultSigner).setUSDLemma(signer1.address);
-        expect(await perpLemma.usdLemma()).to.equal(signer1.address);
 
-        //setReferrer
-        await expect(perpLemma.connect(signer1).setReferrerCode(
-            ethers.utils.formatBytes32String("Hello World")
-        )).to.be.revertedWith("Ownable: caller is not the owner");
+    async function calcLeverage() {
+        positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address)
+        depositedCollateral = await vault.getBalance(perpLemma.address)
+        ethPrice = await mockedBaseAggregator2.getRoundData(0) //ethPrice
+        interval = await clearingHouseConfig.getTwapInterval()
+        usdcPriceInEth = await baseToken.getIndexPrice(interval)
+        leverage_in_6_Decimal = depositedCollateral.mul(parseUnits('1', 26)).div(positionSize.mul(usdcPriceInEth))
+        leverage_in_1 = depositedCollateral.mul(parseUnits('1', 18)).div(positionSize.mul(usdcPriceInEth))
+    
+        // console.log('\nethPrice: ', ethPrice[1].toString())
+        // console.log('usdcPriceInEth: ', usdcPriceInEth.toString())
+        // console.log('positionSize: ', positionSize.toString())
+        // console.log('depositedCollateral: ', depositedCollateral.toString())
+    
+        return [ leverage_in_6_Decimal, leverage_in_1 ]
+    }
 
-        await perpLemma.connect(defaultSigner).setReferrerCode(
-            ethers.utils.formatBytes32String("Hello World")
-        );
+    describe("PerpLemma tests => Open, Close, fees, settlement", () => {
+        before(async function () {
+            // prepare collateral for maker
+            const makerCollateralAmount = parseUnits("1000000", collateralDecimals)
+            await collateral.mint(signer1.address, makerCollateralAmount)
+            await collateral.mint(signer2.address, makerCollateralAmount)
             
-        const byteCode = await perpLemma.referrerCode()
-        expect(ethers.utils.parseBytes32String(byteCode)).to.eq("Hello World")
+            const parsedAmount = parseUnits("100000", collateralDecimals)
+            await collateral.connect(signer1).approve(vault.address, ethers.constants.MaxUint256)
+            await collateral.connect(signer2).approve(vault.address, ethers.constants.MaxUint256)
+    
+            await collateral.mint(longAddress.address, parseUnits("10000000000", collateralDecimals))
+            await collateral.connect(longAddress).approve(vault.address, ethers.constants.MaxUint256)
+            await vault.connect(longAddress).deposit(collateral.address, parseUnits("10000", collateralDecimals))
+    
+            // Deposit into vault
+            await vault.connect(signer2).deposit(collateral.address, parsedAmount)
+            await addLiquidity(
+                clearingHouse, signer2, baseToken.address, 
+                parseEther('10000'), parseEther('100'), -887200, 887200
+            )
+        });
+
+        it("should set addresses correctly", async function () {
+            await expect(perpLemma.connect(signer1).setUSDLemma(signer1.address)).to.be.revertedWith("Ownable: caller is not the owner");
+            await perpLemma.connect(defaultSigner).setUSDLemma(signer1.address);
+            expect(await perpLemma.usdLemma()).to.equal(signer1.address);
+            await expect(perpLemma.connect(signer1).setReferrerCode(ethers.utils.formatBytes32String("Hello World"))).to.be.revertedWith("Ownable: caller is not the owner");
+            await perpLemma.connect(defaultSigner).setReferrerCode(ethers.utils.formatBytes32String("Hello World"));
+            const byteCode = await perpLemma.referrerCode()
+            expect(ethers.utils.parseBytes32String(byteCode)).to.eq("Hello World")
+        });
+    
+        it("should fail to open when max position is reached", async function () {
+            const collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+            await perpLemma.setMaxPosition(parseEther('90'));
+            await collateral.mint(usdLemma.address, collateralAmount.add(1))
+            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount.add(1))
+            await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount)).to.be.revertedWith("max position reached");
+        })
+    
+        it("should close position correctly", async function () {
+            let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+            await collateral.mint(usdLemma.address, collateralAmount)
+    
+            // transfer Collateral to perpLemma
+            await collateral.connect(usdLemma).transfer(perpLemma.address, parseEther('1'))
+            collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+            baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth
+
+            collateralAmount = parseUnits("1", collateralDecimals)
+            // Deposit collateral in eth and Short eth and long usdc 
+            await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                    perpLemma.address,                                                  // Trader
+                    baseToken.address,                                                  // Market --> vUSD
+                    parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                    parseUnits('-980100000000000000', 0),                               // Notional
+                    parseUnits('9900000000000000', 0),                                  // Fee
+                    parseUnits('-990000000000000000', 0),                               // OpenNotional
+                    0,                                                                  // PnlToBeRealized
+                    parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                );
+
+            expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+            expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+            positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+            expect(baseAndQuoteValue[0]).to.eq(positionSize)
+
+            baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize)
+            collateralAmount = baseAndQuoteValue[1].mul(parseEther('1')).div(parseEther('0.99'))
+            // long eth and close position, withdraw collateral
+            await expect(await perpLemma.connect(usdLemma).closeWExactCollateral(collateralAmount))
+                .to.emit(clearingHouse, 'PositionChanged')
+                .withArgs(
+                    perpLemma.address,                                                  // Trader
+                    baseToken.address,                                                  // Market --> vUSD
+                    parseUnits('-97058727412628824788', 0),                             // Position, negative because of short? 
+                    parseUnits('980099999999999999', 0),                                // Notional
+                    parseUnits('9801000000000000', 0),                                  // Fee
+                    parseUnits('-2', 0),                                                // OpenNotional
+                    parseUnits('-19700999999999999', 0),                                                                  // PnlToBeRealized
+                    parseUnits('7922816251426433759433623195', 0)                       // sqrtPriceAfterX96
+                );
+                // .withArgs(
+                //     perpLemma.address,                                                  // Trader
+                //     baseToken.address,                                                  // Market --> vUSD
+                //     parseUnits('-96078723462815946048', 0),                             // Position, negative because of short? 
+                //     parseUnits('970299000000000000', 0),                                // Notional
+                //     parseUnits('9702990000000000', 0),                                  // Fee
+                //     parseUnits('-9996050187121161', 0),                                 // OpenNotional
+                //     parseUnits('-19407939812878839', 0),                                // PnlToBeRealized
+                //     parseUnits('7923592766647236064127145474', 0)                       // sqrtPriceAfterX96
+                // );
+            positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+            expect(positionSize).to.closeTo(BigNumber.from('1000'), BigNumber.from('1000'));
+            expect(await vault.getBalance(perpLemma.address)).to.gt(0) // consider to be fee
+            expect(await collateral.balanceOf(perpLemma.address)).to.be.equal(ZERO);
+        })
+        
+        describe("OpenPosition test1", () => {
+            let collateralToGetBack_1e6, collateralToGetBack_1e18
+            beforeEach(async function () {
+                const collateralAmount = parseEther('1')
+                await collateral.mint(usdLemma.address, collateralAmount)
+                await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount)
+            });
+    
+            it("openPosition => emit event PositionChanged", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+                // Deposit collateral in eth and Short eth and long usdc 
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                        parseUnits('-980100000000000000', 0),                               // Notional
+                        parseUnits('9900000000000000', 0),                                  // Fee
+                        parseUnits('-990000000000000000', 0),                               // OpenNotional
+                        0,                                                                  // PnlToBeRealized
+                        parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                    );
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+            });
+    
+            it("openPosition => leverage should be 1x", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                        parseUnits('-980100000000000000', 0),                               // Notional
+                        parseUnits('9900000000000000', 0),                                  // Fee
+                        parseUnits('-990000000000000000', 0),                               // OpenNotional
+                        0,                                                                  // PnlToBeRealized
+                        parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                    );
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+                const depositedCollateral = await vault.getBalance(perpLemma.address)
+                const ethPrice = await mockedBaseAggregator2.getRoundData(0) //ethPrice
+                const leverage = depositedCollateral.mul(ethPrice[1]).div(positionSize) // 979999(close to 1e6 or 1x)
+                expect(leverage).to.be.closeTo(parseUnits('1', collateralDecimals), parseEther('0.031')); // leverage should be 1x(1e6) or close to 1e6
+            });
+        })
+    
+        describe("OpenPosition tests2", () => {
+            let collateralmintAmount
+            beforeEach(async function () {
+                collateralmintAmount = parseEther('1')
+                await collateral.mint(usdLemma.address, collateralmintAmount)
+                await collateral.connect(usdLemma).transfer(perpLemma.address, collateralmintAmount)
+            });
+            
+            it("openPosition => open position for short and close position for 2 time longs", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                        parseUnits('-980100000000000000', 0),                               // Notional
+                        parseUnits('9900000000000000', 0),                                  // Fee
+                        parseUnits('-990000000000000000', 0),                               // OpenNotional
+                        0,                                                                  // PnlToBeRealized
+                        parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                    );  
+                
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+
+                // #1    
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize.div(2))
+                collateralAmount = baseAndQuoteValue[1].mul(parseEther('1')).div(parseEther('0.99'))
+
+                await expect(perpLemma.connect(usdLemma).closeWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,       
+                        parseUnits('-48529363706314412426', 0),                             // Position, negative because of short? 
+                        parseUnits('492439778913434713', 0),                                // Notional
+                        parseUnits('4924397789134348', 0),                                  // Fee
+                        parseUnits('-495000000000000001', 0),                                 // OpenNotional
+                        parseUnits('-7484618875699634', 0),                                                                  // PnlToBeRealized
+                        parseUnits('7961452674674422230504018035', 0)                                                // Market --> vUSD
+                        // parseUnits('-48041715784969225897', 0),                             // Position, negative because of short? 
+                        // parseUnits('487515381124300366', 0),                                // Notional
+                        // parseUnits('4875153811243004', 0),                                  // Fee
+                        // parseUnits('-499974013723457483', 0),                                 // OpenNotional
+                        // parseUnits('-7385758963485155', 0),                                                                  // PnlToBeRealized
+                        // parseUnits('7961842825662744650554972586', 0)                       // sqrtPriceAfterX96
+                    );
+                
+                // #2
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize.div(2))
+                collateralAmount = baseAndQuoteValue[1].mul(parseEther('1')).div(parseEther('0.99'))
+
+                await expect(perpLemma.connect(usdLemma).closeWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('-48529363706314412363', 0),                             // Position, negative because of short? 
+                        parseUnits('487660221086565286', 0),                                // Notional
+                        parseUnits('4876602210865653', 0),                                  // Fee
+                        parseUnits('-2', 0),                               // OpenNotional
+                        parseUnits('-12216381124300366', 0),                                                                  // PnlToBeRealized
+                        parseUnits('7922816251426433759433623194', 0) 
+                        // parseUnits('-48041727150513801827', 0),                             // Position, negative because of short? 
+                        // parseUnits('482830822831654162', 0),                                // Notional
+                        // parseUnits('4828308228316542', 0),                                  // Fee
+                        // parseUnits('-9947911518245487', 0),                               // OpenNotional
+                        // parseUnits('-12023587601874376', 0),                                                                  // PnlToBeRealized
+                        // parseUnits('7923589026764542382606987206', 0)                       // sqrtPriceAfterX96
+                    );
+
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(positionSize).to.closeTo(BigNumber.from('1000'), BigNumber.from('1000'));
+                expect(await vault.getBalance(perpLemma.address)).to.gt(0) // consider to be fee
+                expect(await collateral.balanceOf(perpLemma.address)).to.be.equal(ZERO);
+            });
+
+            it("openPosition => open position for short and close position for long", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                        parseUnits('-980100000000000000', 0),                               // Notional
+                        parseUnits('9900000000000000', 0),                                  // Fee
+                        parseUnits('-990000000000000000', 0),                               // OpenNotional
+                        0,                                                                  // PnlToBeRealized
+                        parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                    );
+                
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize)
+                collateralAmount = baseAndQuoteValue[1].mul(parseEther('1')).div(parseEther('0.99'))
+
+                await expect(perpLemma.connect(usdLemma).closeWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('-97058727412628824788', 0),                             // Position, negative because of short? 
+                        parseUnits('980099999999999999', 0),                                // Notional
+                        parseUnits('9801000000000000', 0),                                  // Fee
+                        parseUnits('-2', 0),                                                // OpenNotional
+                        parseUnits('-19700999999999999', 0),                                                                  // PnlToBeRealized
+                        parseUnits('7922816251426433759433623195', 0)                       // sqrtPriceAfterX96
+                    );
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(positionSize).to.closeTo(BigNumber.from('1000'), BigNumber.from('1000'));
+                expect(await vault.getBalance(perpLemma.address)).to.gt(0) // consider to be fee
+                expect(await collateral.balanceOf(perpLemma.address)).to.be.equal(ZERO);
+            });
+        })
+    
+        describe("OpenWExactCollateral and CloseWExactCollateral", async function () {
+            
+            it("Basic Open", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+                expect(await collateral.balanceOf(perpLemma.address)).to.equal(collateralAmount);
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                        parseUnits('-980100000000000000', 0),                               // Notional
+                        parseUnits('9900000000000000', 0),                                  // Fee
+                        parseUnits('-990000000000000000', 0),                               // OpenNotional
+                        0,                                                                  // PnlToBeRealized
+                        parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                    );
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+            })
+        
+            it("Basic Open and Close, Checking the lost collateral should be < 5%", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+                const usdLemmaBalance1 = await collateral.balanceOf(usdLemma.address);
+                expect(await collateral.balanceOf(perpLemma.address)).to.equal(collateralAmount);
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                    perpLemma.address,                                                  // Trader
+                    baseToken.address,                                                  // Market --> vUSD
+                    parseUnits('97058727412628824887', 0),                              // Taker Position 
+                    parseUnits('-980100000000000000', 0),                               // Notional
+                    parseUnits('9900000000000000', 0),                                  // Fee
+                    parseUnits('-990000000000000000', 0),                               // OpenNotional
+                    0,                                                                  // PnlToBeRealized
+                    parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                );
+
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize)
+                collateralAmount = baseAndQuoteValue[1].mul(parseEther('1')).div(parseEther('0.99'))
+    
+                await expect(perpLemma.connect(usdLemma).closeWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('-97058727412628824788', 0),                             // Position, negative because of short? 
+                        parseUnits('980099999999999999', 0),                                // Notional
+                        parseUnits('9801000000000000', 0),                                  // Fee
+                        parseUnits('-2', 0),                                                // OpenNotional
+                        parseUnits('-19700999999999999', 0),                                                                  // PnlToBeRealized
+                        parseUnits('7922816251426433759433623195', 0)                       // sqrtPriceAfterX96
+                    );
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(positionSize).to.closeTo(BigNumber.from('1000'), BigNumber.from('1000'));
+                expect(await vault.getBalance(perpLemma.address)).to.gt(0) // consider to be fee
+                expect(await collateral.balanceOf(perpLemma.address)).to.be.equal(ZERO);
+
+                const usdLemmaBalance2 = await collateral.balanceOf(usdLemma.address);
+                const deltaBalance = usdLemmaBalance2 - usdLemmaBalance1;
+                const lostCollateral = collateralAmount - deltaBalance;
+                const percLostCollateral = lostCollateral / collateralAmount;    
+                // Checking the lost collateral is < 5% of the initial amount 
+                expect(collateralAmount - deltaBalance).to.below(collateralAmount*0.05);
+            })
+        })
+    
+        describe("Emergency Settlement", async function () {
+            //let collateralmintAmount, collateralAmount, parsedAmount, leveragedAmount
+            beforeEach(async function () {    
+            });
+    
+            it("Calling Settle() when Market is open should revert", async () => {
+                // By default the market is open
+                await expect(perpLemma.connect(usdLemma).settle()).to.be.revertedWith("CH_MNC");
+            })
+    
+            it("Calling Settle() when Market is paused should revert", async () => {
+                // Pausing the market
+                expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated').withArgs(1);
+                await expect(perpLemma.connect(usdLemma).settle()).to.be.revertedWith("CH_MNC");
+            })
+    
+            it("Calling Settle() when Market is closed should work", async () => {
+                // Pausing the market
+                expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated').withArgs(1);
+                // Closing the market
+                expect(await (baseToken.connect(defaultSigner)["close(uint256)"](1))).to.emit(baseToken, 'StatusUpdated').withArgs(2);
+                await expect(perpLemma.connect(usdLemma).settle()).to.emit(vault, "Withdrawn").withArgs(collateral.address, perpLemma.address, 0);
+            })
+    
+            it("Open a Position and Calling Settle() when Market is closed should work", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                    perpLemma.address,                                                  // Trader
+                    baseToken.address,                                                  // Market --> vUSD
+                    parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                    parseUnits('-980100000000000000', 0),                               // Notional
+                    parseUnits('9900000000000000', 0),                                  // Fee
+                    parseUnits('-990000000000000000', 0),                               // OpenNotional
+                    0,                                                                  // PnlToBeRealized
+                    parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                );
+    
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+
+                expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated').withArgs(1);
+                expect(await (baseToken.connect(defaultSigner)["close(uint256)"](1))).to.emit(baseToken, 'StatusUpdated').withArgs(2);
+                expect(await perpLemma.connect(usdLemma).settle()).to.emit(vault, "Withdrawn").withArgs(
+                    collateral.address, 
+                    perpLemma.address, 
+                    parseUnits("10001", 0)); // 999999
+    
+                // This is not passing as 
+                // Initial Collateral: 100000000000
+                // Actual Collateral: 99901980199
+                // So the Vault has less collateral than when it started
+                //expect(await collateral.balanceOf(vault.address)).to.equal(initialVaultCollateral);
+            })
+    
+            it("Test Settle and Withdraw Collateral for 2 Users", async () => {
+                let collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+                await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount)
+                collateralAmount = collateralAmount.sub(collateralAmount.mul(BigNumber.from('10000')).div(BigNumber.from('1000000')))
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, false, true, collateralAmount) // index0: base/usd, index1: quote/eth    
+                collateralAmount = parseUnits("1", collateralDecimals)
+
+                // 3.2 USDLemma calls PerpLemma Open to open a position at the PerpV2 Clearing House
+                await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount))
+                    .to.emit(clearingHouse, 'PositionChanged').withArgs(
+                        perpLemma.address,                                                  // Trader
+                        baseToken.address,                                                  // Market --> vUSD
+                        parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
+                        parseUnits('-980100000000000000', 0),                               // Notional
+                        parseUnits('9900000000000000', 0),                                  // Fee
+                        parseUnits('-990000000000000000', 0),                               // OpenNotional
+                        0,                                                                  // PnlToBeRealized
+                        parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+                    );
+                expect(await collateral.balanceOf(perpLemma.address)).to.eq(0)
+                expect(await vault.getBalance(perpLemma.address)).to.eq(parseEther('1'))
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                expect(baseAndQuoteValue[0]).to.eq(positionSize)
+
+                // Start with Market Open
+                expect(await baseToken.getStatus()).to.be.equal(0);
+    
+                // Pause Market
+                expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated');
+                expect(await baseToken.callStatic.getStatus()).to.be.equal(1);
+    
+                // Close Market
+                expect(await (baseToken.connect(defaultSigner)["close(uint256)"](1))).to.emit(baseToken, 'StatusUpdated');
+                expect(await baseToken.callStatic.getStatus()).to.be.equal(2);
+    
+                // await perpLemma.connect(usdLemma).settle()
+                expect(await perpLemma.connect(usdLemma).settle()).to.emit(clearingHouse, 'PositionChanged');
+    
+                let collateralPerpLemma = await collateral.balanceOf(perpLemma.address);
+                const c1 = collateralPerpLemma*0.2
+                const c1_1e18 = parseEther( c1.toString() ).div(parseUnits('1', collateralDecimals));
+                await expect(perpLemma.connect(usdLemma).closeWExactCollateral(c1_1e18)).to.emit(collateral, 'Transfer');
+
+                collateralPerpLemma = await collateral.balanceOf(perpLemma.address);
+
+                expect(await collateral.balanceOf(perpLemma.address)).to.not.equal(0);
+    
+                // console.log("Trying to call PerpLemma.close() after market settlement to withdraw the remaining 80% of the initial collateral that is now the 100% of the remaining collateral");
+                const c2 = collateralPerpLemma;
+                const c2_1e18 = parseEther( c2.toString() ).div(parseUnits('1', collateralDecimals));
+                await expect(perpLemma.connect(usdLemma).closeWExactCollateral(c2_1e18)).to.emit(collateral, 'Transfer')
+                expect(await collateral.balanceOf(perpLemma.address)).to.equal(0);
+            })
+        })
+
+        describe("Fees module", async function () {
+            it("should open and deduct fees, and close deduct fees, perpLemma should have fees at the end", async function () {
+                const collateralAmount = parseUnits("100", collateralDecimals) // 6 decimal
+                await collateral.mint(usdLemma.address, collateralAmount)
+    
+                // transfer Collateral to perpLemma
+                await collateral.connect(usdLemma).transfer(perpLemma.address, parseEther('1'))
+                // Deposit collateral in eth and Short eth and long usdc 
+                await perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1')) 
+    
+                getBase = await accountBalance.getBase(perpLemma.address, baseToken.address)
+                getQuote = await accountBalance.getQuote(perpLemma.address, baseToken.address)
+                positionValue = await accountBalance.getTotalPositionValue(perpLemma.address, baseToken.address)
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                ratioforMm1 = await vault.getFreeCollateralByRatio(perpLemma.address, 0)
+                depositedCollateral = await vault.getBalance(perpLemma.address)
+    
+                // console.log('\ngetBase: ', getBase.toString())
+                // console.log('getQuote: ', getQuote.toString())
+                // console.log('positionValue: ', positionValue.toString())
+                // console.log('positionSize: ', positionSize.toString())
+                // console.log('ratioforMm1: ', ratioforMm1.toString())
+                // console.log('depositedCollateral: ', depositedCollateral.toString())
+    
+                // const interval = await clearingHouseConfig.getTwapInterval()
+                // const indexPrice = await baseToken.getIndexPrice(interval) // 
+                const ethPrice = await mockedBaseAggregator2.getRoundData(0) //ethPrice
+                const divisor = positionSize.mul(ethPrice[1]).div(parseEther('1'))
+                const depositedCollateralWith1e18 = depositedCollateral.mul(parseEther('1'))
+                const leverage = depositedCollateralWith1e18.div(divisor) // 979999(close to 1e6 or 1x)
+    
+                // console.log('indexPrice: ', indexPrice.toString())
+                // console.log('ethPrice: ', ethPrice[1].toString())
+                // console.log('divisor: ', divisor.toString())
+                console.log('leverage: ', leverage.toString())
+    
+    
+                fee1 = (
+                    await clearingHouse.connect(signer2).callStatic.removeLiquidity({
+                        baseToken: baseToken.address,
+                        lowerTick: -887200, //50000,
+                        upperTick: 887200, //50400,
+                        liquidity: 0,
+                        minBase: 0,
+                        minQuote: 0,
+                        deadline: ethers.constants.MaxUint256,
+                    })
+                ).fee
+                // console.log('\nfee1: ', fee1.toString())
+                expect(fee1).to.be.gt(0) // (> 0)
+    
+                baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize)
+    
+                // long eth and close position, withdraw collateral
+                await perpLemma.connect(usdLemma).closeWExactCollateral(baseAndQuoteValue[1])
+    
+                // after close
+                fee2 = (
+                    await clearingHouse.connect(signer2).callStatic.removeLiquidity({
+                        baseToken: baseToken.address,
+                        lowerTick: -887200, //50000,
+                        upperTick: 887200, //50400,
+                        liquidity: 0,
+                        minBase: 0,
+                        minQuote: 0,
+                        deadline: ethers.constants.MaxUint256,
+                    })
+                ).fee
+                // console.log('fee2: ', fee2.toString())      
+                expect(fee2).to.be.gt(0) // (> 0)
+    
+                getBase = await accountBalance.getBase(perpLemma.address, baseToken.address)
+                getQuote = await accountBalance.getQuote(perpLemma.address, baseToken.address)
+                positionValue = await accountBalance.getTotalPositionValue(perpLemma.address, baseToken.address)
+                positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
+                ratioforMm1 = await vault.getFreeCollateralByRatio(perpLemma.address, 0)
+                depositedCollateral = await vault.getBalance(perpLemma.address)
+    
+                // console.log('\ngetBase: ', getBase.toString())
+                // console.log('getQuote: ', getQuote.toString())
+                // console.log('positionValue: ', positionValue.toString())
+                // console.log('positionSize: ', positionSize.toString())
+                // console.log('ratioforMm1: ', ratioforMm1.toString())
+                // console.log('depositedCollateral: ', depositedCollateral.toString())
+    
+                const perpBalance = await collateral.balanceOf(perpLemma.address)
+                expect(perpBalance).to.be.eq(0) // (= 0) Fees charged by perplemma
+    
+                expect(getBase).to.be.closeTo(parseEther('1'), parseEther('0.1'))
+                expect(getQuote.mul(-1)).to.be.closeTo(parseEther('0.0099'), parseEther('0.0009'))
+                expect(positionSize).to.be.closeTo(parseEther('1'), parseEther('0.1'))
+            })    
+        })
     });
 
-    it("should fail to open when max position is reached", async function () {
-        const collateralAmount = parseUnits("1", collateralDecimals) // 6 decimal
-        await perpLemma.setMaxPosition(collateralAmount);
-        await collateral.mint(usdLemma.address, collateralAmount.add(1))
-        // await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount.add(1))
-        // await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralAmount.add(1))).to.be.revertedWith("max position reached");
-    })
+    describe("Rebalance Tests", () => {
+        before(async function () {
+            await perpLemma.connect(defaultSigner).setReBalancer(reBalancer.address);
+            await addLiquidity(
+                clearingHouse, signer2, baseToken.address, 
+                parseEther('100000'), parseEther('1000'), -887200, 887200
+            )
 
-    // need to correct more for collateralAmountToGetBack in close() position
-    it("should close position correctly", async function () {
-        const collateralAmount = parseUnits("100", collateralDecimals) // 6 decimal
-        await collateral.mint(usdLemma.address, collateralAmount)
+            // alice add long limit order
+            await collateral.mint(signer1.address, parseUnits("10000000000", collateralDecimals))
+            await collateral.connect(signer1).approve(vault.address, ethers.constants.MaxUint256)
+            await vault.connect(signer1).deposit(collateral.address, parseUnits("10000", collateralDecimals))
 
-        // transfer Collateral to perpLemma
-        await collateral.connect(usdLemma).transfer(perpLemma.address, parseEther('1'))
-        // Deposit collateral in eth and Short eth and long usdc 
-        await perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1'))
+            await collateral.mint(signer2.address, parseUnits("10000000000", collateralDecimals))
+            await collateral.connect(signer2).approve(vault.address, ethers.constants.MaxUint256)
+            await vault.connect(signer2).deposit(collateral.address, parseUnits("100000", collateralDecimals))
 
-        positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
-        // long eth and close position, withdraw collateral
+            await collateral.mint(longAddress.address, parseUnits("10000000000", collateralDecimals))
+            await collateral.connect(longAddress).approve(vault.address, ethers.constants.MaxUint256)
+            await vault.connect(longAddress).deposit(collateral.address, parseUnits("10000", collateralDecimals))
 
-        baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize)
-        await expect(await perpLemma.connect(usdLemma)
-            .closeWExactCollateral(baseAndQuoteValue[1])).to.emit(clearingHouse, 'PositionChanged')
-
-        positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
-        // console.log('positionSize: ', positionSize.toString())
-        depositedCollateral = await vault.getBalance(perpLemma.address)
-        // console.log('depositedCollateral: ', depositedCollateral.toString())
-    })
-    
-    describe("OpenPosition", () => {
-        let collateralToGetBack_1e6, collateralToGetBack_1e18
-        beforeEach(async function () {
-
-            const collateralAmount = parseEther('1')
-            await collateral.mint(usdLemma.address, collateralAmount)
-            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralAmount)
+            await collateral.mint(usdLemma.address, parseUnits("10000000000", collateralDecimals))
+            await collateral.connect(usdLemma).approve(vault.address, ethers.constants.MaxUint256)
+            await vault.connect(usdLemma).deposit(collateral.address, parseUnits("10000", collateralDecimals))
         });
 
-        it("openPosition => emit event PositionChanged", async () => {
-            await expect(perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1'))).to.emit(clearingHouse, 'PositionChanged')            
-        });
-
-        it("openPosition => leverage should be 1x", async () => {
-            await expect(perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1'))).to.emit(clearingHouse, 'PositionChanged')            
-            const depositedCollateral = await vault.getBalance(perpLemma.address)
-            const positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address)
-            const ethPrice = await mockedBaseAggregator2.getRoundData(0) //ethPrice
-            const leverage = depositedCollateral.mul(ethPrice[1]).div(positionSize) // 979999(close to 1e6 or 1x)
-            // console.log('positionSize:' , positionSize.toString())
-            // console.log('ethPrice:' , ethPrice[1].toString())
-            // console.log('divisor:' , divisor.toString())
-            // console.log('depositedCollateral:' , depositedCollateral.toString())
-            // console.log('leverage:' , leverage.toString())
-            expect(leverage).to.be.closeTo(parseUnits('1', collateralDecimals), parseEther('0.031')); // leverage should be 1x(1e6) or close to 1e6
-        });
-    })
-
-    describe("OpenPosition with getCollateralAmountGivenUnderlyingAssetAmount", () => {
-        let collateralmintAmount
-        beforeEach(async function () {
-            collateralmintAmount = parseEther('1')
-            await collateral.mint(usdLemma.address, collateralmintAmount)
-        });
-
-        it("openPosition => open position for short and close position for 2 time longs", async () => {
-            // let collateralToGetBack_1e18 = await perpLemma.callStatic.getCollateralAmountGivenUnderlyingAssetAmount(parseEther('1'), true)
-            // let collateralToGetBack_1e6 = collateralToGetBack_1e18.mul(parseUnits('1', collateralDecimals)).div(parseEther('1'))
-            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralmintAmount)
-            await perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1'))
-
-            let positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address)
-            
-            // #1    
-            baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize.div(2))
-            await expect(perpLemma.connect(usdLemma).closeWExactCollateral(baseAndQuoteValue[1])).to.emit(clearingHouse, 'PositionChanged')
-            
-            // #2
-            baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize.div(2))
-            await expect(perpLemma.connect(usdLemma).closeWExactCollateral(baseAndQuoteValue[1])).to.emit(clearingHouse, 'PositionChanged')
-            positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
-            expect(positionSize).to.be.closeTo(parseEther('1'), parseEther('0.1'))
-        });
-
-        it("openPosition => open position for short and close position for long", async () => {
-            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralmintAmount)
-            await perpLemma.connect(usdLemma).openWExactCollateral(collateralmintAmount)
-            let positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address)
-            baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, positionSize)
-            await expect(perpLemma.connect(usdLemma).closeWExactCollateral(baseAndQuoteValue[1])).to.emit(clearingHouse, 'PositionChanged')
-            positionSize = await accountBalance.getTotalPositionSize(perpLemma.address, baseToken.address) 
-            expect(positionSize).to.be.closeTo(parseEther('1'), parseEther('0.1'))
-        });
-    })
-
-    // describe("re balance", async function () {
-    //     let collateralmintAmount, collateralAmount, parsedAmount, leveragedAmount
-    //     before(async function () {
-    //         await perpLemma.connect(defaultSigner).setReBalancer(reBalancer.address);
-    //     })
-    //     beforeEach(async function () {
-    //         collateralmintAmount = parseEther('1')
-    //         // collateralAmount = parseUnits("1000", collateralDecimals) // 6 decimal
-    //         // parsedAmount =  collateralAmount.mul(parseEther('1')).div(parseUnits('1', 6)) // 18 decimal
-    //         // leveragedAmount = parsedAmount.mul('1') // for 1x
-    //         await collateral.mint(usdLemma.address, collateralmintAmount)
-    //     });
-
-    //     it("if amount is positive then it should long", async () => {
-    //         const sqrtPriceLimitX96 = 0;
-    //         const deadline = ethers.constants.MaxUint256;
-    //         await collateral.connect(usdLemma).transfer(perpLemma.address, collateralmintAmount)
-    //         await perpLemma.connect(usdLemma).openWExactCollateral(collateralmintAmount)
-    //         const rebalanceAmount = parseEther('0.05') // 5% of 1 ether
-    //         await perpLemma.connect(usdLemma).reBalance(
-    //             reBalancer.address, 
-    //             BigNumber.from(rebalanceAmount), // positive amount(+ve)
-    //             ethers.utils.defaultAbiCoder.encode(
-    //                 ["uint160", "uint256"], 
-    //                 [sqrtPriceLimitX96, deadline]
-    //             )
-    //         );
-    //     })
-
-    //     it("if amount is negative then it should short", async () => {
-    //         const sqrtPriceLimitX96 = 0;
-    //         const deadline = ethers.constants.MaxUint256;
-    //         // await expect(perpLemma.connect(usdLemma).open(leveragedAmount, collateralAmount)).to.emit(clearingHouse, 'PositionChanged')
-    //         await collateral.connect(usdLemma).transfer(perpLemma.address, collateralmintAmount)
-    //         await perpLemma.connect(usdLemma).openWExactCollateral(collateralmintAmount)
-    //         const rebalanceAmount = parseEther('0.05') // 5% of 1 ether
-    //         await perpLemma.connect(usdLemma).reBalance(
-    //             reBalancer.address, 
-    //             BigNumber.from(rebalanceAmount).mul(-1), // negative amount(-ve)
-    //             ethers.utils.defaultAbiCoder.encode(
-    //                 ["uint160", "uint256"], 
-    //                 [sqrtPriceLimitX96, deadline]
-    //             )
-    //         );
-    //     })
-    // })
-   
-    describe("OpenWExactCollateral and CloseWExactCollateral", async function () {
-        
-        it("Basic Open", async () => {
-            const collateralAmount = parseUnits("100", collateralDecimals) // 6 decimal
-            await collateral.mint(usdLemma.address, collateralAmount);
-
-            // Open a Position
-            // getCollateralAmountGivenUnderlyingAssetAmount() DEPRECATED --> Let's replace with a fixed Collateral Amount 1 ETH position
-            //const desiredAmountUSDL = parseEther('1');
-            collateralRequired_1e18 = parseEther('1');
-            //collateralRequired_1e18 = await perpLemma.callStatic.getCollateralAmountGivenUnderlyingAssetAmount(desiredAmountUSDL, true)
-            collateralRequired_1eCollateralDecimals = collateralRequired_1e18.mul(parseUnits('1', collateralDecimals)).div(parseEther('1'))
-
-            console.log(`T1 Collateral Required in 1e18 = ${collateralRequired_1e18} and in 1e${collateralDecimals} = ${collateralRequired_1eCollateralDecimals}`);
-            
-            // We need to convety
-            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralRequired_1eCollateralDecimals);
-            expect(await collateral.balanceOf(perpLemma.address)).to.equal(collateralRequired_1eCollateralDecimals);
-
-            
-            await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralRequired_1e18)).to.emit(clearingHouse, 'PositionChanged').withArgs(
-                perpLemma.address,                                                  // Trader
-                baseToken.address,                                                  // Market --> vUSD
-                parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
-                parseUnits('-980100000000000000', 0),                               // Notional
-                parseUnits('9900000000000000', 0),                                  // Fee
-                parseUnits('-990000000000000000', 0),                               // OpenNotional
-                0,                                                                  // PnlToBeRealized
-                parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+        it("#1.a Rebalance, only usdLemma with short eth", async () => {
+            await collateral.connect(usdLemma).transfer(perpLemma.address, parseEther('1'))
+            await perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1')) 
+            await forwardTimestamp(clearingHouse, 200)
+            await clearingHouse.settleAllFunding(perpLemma.address)
+            await forwardTimestamp(clearingHouse, 200)
+            let fundingPayment = await exchange.getPendingFundingPayment(perpLemma.address, baseToken.address)
+            console.log('fundingPayment: ', fundingPayment.toString())
+            let leverage_before  = await calcLeverage()
+            console.log('leverage_before_in_6_decimal: ', leverage_before[0].toString())
+            console.log('leverage_before_in_1: ', leverage_before[1].toString())
+            fundingPNL = await perpLemma.getFundingPNL()
+            realizedFundingPnl = await perpLemma.realizedFundingPNL()
+            console.log('fundingPNL: ', fundingPNL.toString())
+            console.log('realizedFundingPnl: ', realizedFundingPnl.toString())
+            const sqrtPriceLimitX96 = 0;
+            const deadline = ethers.constants.MaxUint256;
+            await perpLemma.connect(usdLemma).reBalance(
+                reBalancer.address,
+                fundingPNL.sub(realizedFundingPnl),
+                ethers.utils.defaultAbiCoder.encode(
+                    ["uint160", "uint256"],
+                    [sqrtPriceLimitX96, deadline]
+                )
             );
-
-            expect(await collateral.balanceOf(perpLemma.address)).to.equal(0);
-            
+            let leverage_after  = await calcLeverage()
+            console.log('leverage_after_in_6_decimal: ', leverage_after[0].toString())
+            console.log('leverage_after_in_1: ', leverage_after[1].toString())
+            expect(leverage_after[1]).eq(1)
         })
-    
 
-        
-        it("Basic Open and Close", async () => {
-            const collateralAmount = parseUnits("100", collateralDecimals) // 6 decimal
-            await collateral.mint(usdLemma.address, collateralAmount);
-
-            // Open a Position
-            // getCollateralAmountGivenUnderlyingAssetAmount() DEPRECATED --> Let's replace with a fixed Collateral Amount 1 ETH position
-            //const desiredAmountUSDL = parseEther('1');
-            collateralRequired_1e18 = parseEther('1');
-            //collateralRequired_1e18 = await perpLemma.callStatic.getCollateralAmountGivenUnderlyingAssetAmount(desiredAmountUSDL, true)
-            collateralRequired_1eCollateralDecimals = collateralRequired_1e18.mul(parseUnits('1', collateralDecimals)).div(parseEther('1'))
-
-            console.log(`T1 Collateral Required in 1e18 = ${collateralRequired_1e18} and in 1e${collateralDecimals} = ${collateralRequired_1eCollateralDecimals}`);
-            // We need to convety
-            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralRequired_1eCollateralDecimals);
-            const usdLemmaBalance1 = await collateral.balanceOf(usdLemma.address);
-            expect(await collateral.balanceOf(perpLemma.address)).to.equal(collateralRequired_1eCollateralDecimals);
-
-            await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralRequired_1e18)).to.emit(clearingHouse, 'PositionChanged').withArgs(
-                perpLemma.address,                                                  // Trader
-                baseToken.address,                                                  // Market --> vUSD
-                parseUnits('97058727412628824887', 0),                              // Taker Position 
-                parseUnits('-980100000000000000', 0),                               // Notional
-                parseUnits('9900000000000000', 0),                                  // Fee
-                parseUnits('-990000000000000000', 0),                               // OpenNotional
-                0,                                                                  // PnlToBeRealized
-                parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+        it("#1.b Rebalance, longAddress(signer) with long eth, only usdLemma short eth", async () => {
+            await openPosition(clearingHouse, longAddress, baseToken.address, true, false, parseEther('8'))
+            await collateral.connect(usdLemma).transfer(perpLemma.address, parseEther('1'))
+            await perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1')) 
+            await forwardTimestamp(clearingHouse, 200)
+            await clearingHouse.settleAllFunding(perpLemma.address)
+            await forwardTimestamp(clearingHouse, 200)
+            let fundingPayment = await exchange.getPendingFundingPayment(perpLemma.address, baseToken.address)
+            console.log('fundingPayment: ', fundingPayment.toString())
+            let leverage_before  = await calcLeverage()
+            console.log('leverage_before_in_6_decimal: ', leverage_before[0].toString())
+            console.log('leverage_before_in_1: ', leverage_before[1].toString())
+            fundingPNL = await perpLemma.getFundingPNL()
+            realizedFundingPnl = await perpLemma.realizedFundingPNL()
+            totalFundingPNL = await perpLemma.totalFundingPNL()
+            console.log('fundingPNL: ', fundingPNL.toString())
+            console.log('realizedFundingPnl: ', realizedFundingPnl.toString())
+            const sqrtPriceLimitX96 = 0;
+            const deadline = ethers.constants.MaxUint256;
+            await perpLemma.connect(usdLemma).reBalance(
+                reBalancer.address,
+                fundingPNL.sub(realizedFundingPnl),
+                ethers.utils.defaultAbiCoder.encode(
+                    ["uint160", "uint256"],
+                    [sqrtPriceLimitX96, deadline]
+                )
             );
+            let leverage_after  = await calcLeverage()
+            console.log('leverage_after_in_6_decimal: ', leverage_after[0].toString())
+            console.log('leverage_after_in_1: ', leverage_after[1].toString())
 
-            expect(await collateral.balanceOf(perpLemma.address)).to.equal(0);
-            const takerPositionSize = await accountBalance.getTakerPositionSize(perpLemma.address, baseToken.address);
-            baseAndQuoteValue = await callStaticOpenPosition(clearingHouse, longAddress, baseToken.address, true, true, takerPositionSize)
-
-            expect(parseUnits(takerPositionSize.toString(), 0)).to.equal(parseUnits('97058727412628824887', 0));
-            await expect(perpLemma.connect(usdLemma).closeWExactCollateral(baseAndQuoteValue[1])).to.emit(collateral, 'Transfer');
-
-            const usdLemmaBalance2 = await collateral.balanceOf(usdLemma.address);
-            const deltaBalance = usdLemmaBalance2 - usdLemmaBalance1;
-            const lostCollateral = collateralRequired_1eCollateralDecimals - deltaBalance;
-            const percLostCollateral = lostCollateral / collateralRequired_1eCollateralDecimals;
-            console.log(`Spent Collateral = ${collateralRequired_1eCollateralDecimals}, Recovered Collateral = ${deltaBalance}, Lost Collateral = ${lostCollateral}, percLostCollateral = ${percLostCollateral}`);
-
-            // Checking the lost collateral is < 5% of the initial amount 
-            expect(collateralRequired_1eCollateralDecimals - deltaBalance).to.below(collateralRequired_1eCollateralDecimals*0.05);
-        })
-        
-
-    })
-
-    
-    describe("Emergency Settlement", async function () {
-        //let collateralmintAmount, collateralAmount, parsedAmount, leveragedAmount
-        beforeEach(async function () {
-
-        });
-
-        it("Calling Settle() when Market is open should revert", async () => {
-            // By default the market is open
-            await expect(perpLemma.connect(usdLemma).settle()).to.be.revertedWith("CH_MNC");
+            expect(leverage_after[0]).lt(leverage_before[0])
+            expect(leverage_after[1]).eq(1)
         })
 
-        it("Calling Settle() when Market is paused should revert", async () => {
-            // Pausing the market
-            expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated').withArgs(1);
-            await expect(perpLemma.connect(usdLemma).settle()).to.be.revertedWith("CH_MNC");
-        })
-
-        it("Calling Settle() when Market is closed should work", async () => {
-            // Pausing the market
-            expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated').withArgs(1);
-            // Closing the market
-            expect(await (baseToken.connect(defaultSigner)["close(uint256)"](1))).to.emit(baseToken, 'StatusUpdated').withArgs(2);
-            await expect(perpLemma.connect(usdLemma).settle()).to.emit(vault, "Withdrawn").withArgs(collateral.address, perpLemma.address, 0);
-        })
-
-        it("Open a Position and Calling Settle() when Market is closed should work", async () => {
-            const collateralAmount = parseUnits("100", collateralDecimals) // 6 decimal
-            await collateral.mint(usdLemma.address, collateralAmount)
-
-            // getCollateralAmountGivenUnderlyingAssetAmount() DEPRECATED --> Let's replace with a fixed Collateral Amount 1 ETH position
-            //const desiredAmountUSDL = parseEther('1');
-            collateralRequired_1e18 = parseEther('1');
-            //collateralRequired_1e18 = await perpLemma.callStatic.getCollateralAmountGivenUnderlyingAssetAmount(desiredAmountUSDL, true)
-            collateralRequired_1e6 = collateralRequired_1e18.mul(parseUnits('1', collateralDecimals)).div(parseEther('1'))
-
-            console.log(`T1 Collateral Required in 1e18 = ${collateralRequired_1e18} and in 1e${collateralDecimals} = ${collateralRequired_1e6}`);
-            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralRequired_1e6);
-            expect(await collateral.balanceOf(perpLemma.address)).to.equal(collateralRequired_1e6);
-
-            // Initially the Vault should have no collateral
-            const initialVaultCollateral = await collateral.balanceOf(vault.address);
-            //expect(await collateral.balanceOf(vault.address)).to.equal(0);
-
-            await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralRequired_1e18)).to.emit(clearingHouse, 'PositionChanged').withArgs(
-                perpLemma.address,                                                  // Trader
-                baseToken.address,                                                  // Market --> vUSD
-                parseUnits('97058727412628824887', 0),                              // Position, negative because of short? 
-                parseUnits('-980100000000000000', 0),                               // Notional
-                parseUnits('9900000000000000', 0),                                  // Fee
-                parseUnits('-990000000000000000', 0),                               // OpenNotional
-                0,                                                                  // PnlToBeRealized
-                parseUnits('8000467773506664236629439201', 0)                       // sqrtPriceAfterX96
+        it("#1.c Rebalance, longAddress(signer) with long eth, signer1(signer) with short eth, and usdLemma short eth", async () => {
+            await openPosition(clearingHouse, longAddress, baseToken.address, true, false, parseEther('8'))
+            await openPosition(clearingHouse, signer1, baseToken.address, false, true, parseEther('5'))
+            await collateral.connect(usdLemma).transfer(perpLemma.address, parseEther('1'))
+            await perpLemma.connect(usdLemma).openWExactCollateral(parseEther('1')) 
+            await forwardTimestamp(clearingHouse, 200)
+            await clearingHouse.settleAllFunding(perpLemma.address)
+            await forwardTimestamp(clearingHouse, 200)
+            let fundingPayment = await exchange.getPendingFundingPayment(perpLemma.address, baseToken.address)
+            console.log('fundingPayment: ', fundingPayment.toString())
+            let leverage_before  = await calcLeverage()
+            console.log('leverage_before_in_6_decimal: ', leverage_before[0].toString())
+            console.log('leverage_before_in_1: ', leverage_before[1].toString())
+            fundingPNL = await perpLemma.getFundingPNL()
+            realizedFundingPnl = await perpLemma.realizedFundingPNL()
+            totalFundingPNL = await perpLemma.totalFundingPNL()
+            console.log('fundingPNL: ', fundingPNL.toString())
+            console.log('realizedFundingPnl: ', realizedFundingPnl.toString())
+            const sqrtPriceLimitX96 = 0;
+            const deadline = ethers.constants.MaxUint256;
+            await perpLemma.connect(usdLemma).reBalance(
+                reBalancer.address,
+                fundingPNL.sub(realizedFundingPnl),
+                ethers.utils.defaultAbiCoder.encode(
+                    ["uint160", "uint256"],
+                    [sqrtPriceLimitX96, deadline]
+                )
             );
-
-            // All the collateral computed with `getCollateralAmountGivenUnderlyingAssetAmount()` is transferred to PerpLemma that deposits all in the Vault to Open a Position 
-            expect(await collateral.balanceOf(perpLemma.address)).to.equal(0);
-
-            // Pausing the market
-            expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated').withArgs(1);
-            // Closing the market
-            expect(await (baseToken.connect(defaultSigner)["close(uint256)"](1))).to.emit(baseToken, 'StatusUpdated').withArgs(2);
-            expect(await perpLemma.connect(usdLemma).settle()).to.emit(vault, "Withdrawn").withArgs(
-                collateral.address, 
-                perpLemma.address, 
-                parseUnits("10001", 0)); // 999999
-
-            // This is not passing as 
-            // Initial Collateral: 100000000000
-            // Actual Collateral: 99901980199
-            // So the Vault has less collateral than when it started
-            //expect(await collateral.balanceOf(vault.address)).to.equal(initialVaultCollateral);
-        })
-
-        it("Test Settle and Withdraw Collateral for 2 Users", async () => {
-            // 1. Mint
-            const collateralAmount = parseUnits("100", collateralDecimals) // 6 decimal
-            await collateral.mint(usdLemma.address, collateralAmount)
-            
-            const collateralUSDLemma_t0 = await collateral.balanceOf(usdLemma.address);
-            const collateralPerpLemma_t0 = await collateral.balanceOf(perpLemma.address);
-
-            console.log("1. Initial Balances");
-            console.log(`1. USDLemma Balance = ${collateralUSDLemma_t0}`); 
-            console.log(`1. PerpLemma Balance = ${collateralPerpLemma_t0}`);
-
-            // 2. Get amount of collateral
-            // getCollateralAmountGivenUnderlyingAssetAmount() DEPRECATED --> Let's replace with a fixed Collateral Amount 1 ETH position
-            //const desiredAmountUSDL = parseEther('1');
-            collateralRequired_1e18 = parseEther('1');
-            //collateralRequired_1e18 = await perpLemma.callStatic.getCollateralAmountGivenUnderlyingAssetAmount(desiredAmountUSDL, true)
-            collateralRequired_1e6 = collateralRequired_1e18.mul(parseUnits('1', collateralDecimals)).div(parseEther('1'))
-            //console.log(`Collateral Required to Open a short ${perpPosition} (1e18) on BaseToken (vUSD) = ${collateralRequired_1e6} (1e6) Collateral (ETH)`);
-
-            // 3. Open Position
-            // 3.1 Transfer from USDLemma (High Level Abstraction Trader)  --> PerpLemma (Backend Protocol Specific Trader)
-            await collateral.connect(usdLemma).transfer(perpLemma.address, collateralRequired_1e6);
-
-            const collateralUSDLemma_t1 = await collateral.balanceOf(usdLemma.address);
-            const collateralPerpLemma_t1 = await collateral.balanceOf(perpLemma.address);
-
-            console.log("2. Balances after transfer");
-            console.log(`2. USDLemma Balance = ${collateralUSDLemma_t1}, delta = ${collateralUSDLemma_t1 - collateralUSDLemma_t0}`); 
-            console.log(`2. PerpLemma Balance = ${collateralPerpLemma_t1}, delta = ${collateralPerpLemma_t1 - collateralPerpLemma_t0}`);
-
-            // 3.2 USDLemma calls PerpLemma Open to open a position at the PerpV2 Clearing House
-            await expect(perpLemma.connect(usdLemma).openWExactCollateral(collateralRequired_1e18)).to.emit(clearingHouse, 'PositionChanged');
-
-            const collateralUSDLemma_t2 = await collateral.balanceOf(usdLemma.address);
-            const collateralPerpLemma_t2 = await collateral.balanceOf(perpLemma.address);
-
-            console.log("3. Balances after Open");
-            console.log(`3. USDLemma Balance = ${collateralUSDLemma_t2}, delta = ${collateralUSDLemma_t2 - collateralUSDLemma_t1}`); 
-            console.log(`3. PerpLemma Balance = ${collateralPerpLemma_t2}, delta = ${collateralPerpLemma_t2 - collateralPerpLemma_t1}`);
-
-            // Start with Market Open
-            expect(await baseToken.getStatus()).to.be.equal(0);
-
-            // Pause Market
-            expect(await (baseToken.connect(defaultSigner)["pause(uint256)"](0))).to.emit(baseToken, 'StatusUpdated');
-            expect(await baseToken.callStatic.getStatus()).to.be.equal(1);
-
-            // Close Market
-            expect(await (baseToken.connect(defaultSigner)["close(uint256)"](1))).to.emit(baseToken, 'StatusUpdated');
-            expect(await baseToken.callStatic.getStatus()).to.be.equal(2);
-
-            // await perpLemma.connect(usdLemma).settle()
-            expect(await perpLemma.connect(usdLemma).settle()).to.emit(clearingHouse, 'PositionChanged');
-
-            const collateralUSDLemma_t32 = await collateral.balanceOf(usdLemma.address);
-            const collateralPerpLemma_t32 = await collateral.balanceOf(perpLemma.address);
-
-            console.log("5. Balances after Settlement");
-            console.log(`5. USDLemma Balance = ${collateralUSDLemma_t32}, delta = ${collateralUSDLemma_t32 - collateralUSDLemma_t2}`); 
-            console.log(`5. PerpLemma Balance = ${collateralPerpLemma_t32}, delta = ${collateralPerpLemma_t32 - collateralPerpLemma_t2}, delta with initial ${collateralPerpLemma_t32 - collateralPerpLemma_t1}`);
-            
-            console.log(`5. PositionAtSettlement = ${await perpLemma.positionAtSettlement()}`);
-
-            console.log("Trying to call PerpLemma.close() after market settlement to withdraw 20% of collateral");
-            // No need to specify the expected collateral amount when the market is closed, it is computed as a percentage of the positionAtSettlement
-            const c1 = collateralPerpLemma_t32*0.2
-            const c1_1e18 = parseEther( c1.toString() ).div(parseUnits('1', collateralDecimals));
-            console.log(`c1_1e18 = ${c1_1e18}`);
-            await expect(perpLemma.connect(usdLemma).closeWExactCollateral(c1_1e18)).to.emit(collateral, 'Transfer');
-            console.log("Closed 20% DONE");
-
-            const collateralUSDLemma_t33 = await collateral.balanceOf(usdLemma.address);
-            const collateralPerpLemma_t33 = await collateral.balanceOf(perpLemma.address);
-
-            console.log("Balances after transfer");
-            console.log(`USDLemma Balance = ${collateralUSDLemma_t33}, delta = ${collateralUSDLemma_t33 - collateralUSDLemma_t32}`); 
-            console.log(`PerpLemma Balance = ${collateralPerpLemma_t33}, delta = ${collateralPerpLemma_t33 - collateralPerpLemma_t32}`);
-
-            expect(await collateral.balanceOf(perpLemma.address)).to.not.equal(0);
-
-            console.log("Trying to call PerpLemma.close() after market settlement to withdraw the remaining 80% of the initial collateral that is now the 100% of the remaining collateral");
-            const c2 = collateralPerpLemma_t33;
-            const c2_1e18 = parseEther( c2.toString() ).div(parseUnits('1', collateralDecimals));
-            await expect(perpLemma.connect(usdLemma).closeWExactCollateral(c2_1e18)).to.emit(collateral, 'Transfer');
-
-            const collateralUSDLemma_t35 = await collateral.balanceOf(usdLemma.address);
-            const collateralPerpLemma_t35 = await collateral.balanceOf(perpLemma.address);
-
-            console.log("Balances after transfer");
-            console.log(`USDLemma Balance = ${collateralUSDLemma_t35}, delta = ${collateralUSDLemma_t35 - collateralUSDLemma_t33}`); 
-            console.log(`PerpLemma Balance = ${collateralPerpLemma_t35}, delta = ${collateralPerpLemma_t35 - collateralPerpLemma_t33}`);
-
-            expect(await collateral.balanceOf(perpLemma.address)).to.equal(0);
+            let leverage_after  = await calcLeverage()
+            console.log('leverage_after_in_6_decimal: ', leverage_after[0].toString())
+            console.log('leverage_after_in_1: ', leverage_after[1].toString())
+            expect(leverage_after[0]).lt(leverage_before[0])
+            expect(leverage_after[1]).eq(1)
         })
     })
-
 })
